+++ cli.js
@@ -2,7 +2,8 @@
 const { program } = require('commander');
 program.version(require('./package.json').version);
 
-const { create, refund, upload, remove, setDefault, download } = require("./index");
+const { create, refund, remove, setDefault, download } = require("./index");
+const { uploadEvent } =  require("./src/events");
 
 program
     .option('-p, --privateKey [privateKey]', 'private key')
@@ -86,7 +87,7 @@ program
     .option('-g, --gasPriceIncreasePercentage [gasPriceIncreasePercentage]', 'gas price increase percentage')
     .action(() => {
         const opts = program.opts();
-        upload(opts.privateKey, opts.address, opts.file, opts.type, opts.rpc,  opts.chainId, opts.gasPriceIncreasePercentage);
+        uploadEvent(opts.privateKey, opts.address, opts.file, opts.type, opts.rpc, opts.chainId, opts.gasPriceIncreasePercentage);
     });
 
 program.parse(process.argv);
