+++ package.json
@@ -14,6 +14,7 @@
   "author": "ethstorage",
   "license": "MIT",
   "dependencies": {
+    "@inquirer/prompts": "^5.0.2",
     "colors-cli": "^1.0.33",
     "commander": "^11.1.0",
     "eth-ens-namehash": "^2.0.8",
@@ -21,6 +22,7 @@
     "ethstorage-sdk": "^1.2.8",
     "fs": "^0.0.1-security",
     "js-sha3": "^0.8.0",
+    "ora": "^8.0.1",
     "rxjs": "^7.8.1"
   },
   "engines": {
